requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
ansicon==1.89.0
arrow==1.2.3
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
black==23.1.0
blessed==1.20.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cron-descriptor==1.2.35
Django==4.1.7
django-picklefield==3.1
django-q==1.3.9
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.1
frozenlist==1.3.3
idna==3.4
inflection==0.5.1
jinxed==1.2.0
jsonschema==4.17.3
kombu==5.2.4
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
prompt-toolkit==3.0.38
PyJWT==2.6.0
pyrsistent==0.19.3
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.2
PyYAML==6.0
redis==3.5.3
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
stripe==5.3.0
tomli==2.0.1
tzdata==2023.2
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
yarl==1.8.2
